Change Double Quotes Behavior
In order to use the SQL Parser for Athena SQL, we need to change the behavior of tokens that are enclosed in double quotes. Instead of treating them as strings like in MySQL, treat them as identifiers (column and table names) like in Athena.
In addition, the SQL Parser is not allowed to include tokens that are enclosed in backticks now. I did it for reasons of code simplicity and the fact that backtricks are used to escape reserved keywords in DDL statements and our parser isn't supposed to parse DDL statements.